add money saving goal
saving request amount -> bigdecimal

+++ app/src/main/java/com/example/happy_wallet_mobile/Data/Remote/Request/SavingGoal/CreateSavingGoalRequest.java
@@ -1,16 +1,18 @@
 package com.example.happy_wallet_mobile.Data.Remote.Request.SavingGoal;
 
+import java.math.BigDecimal;
+
 public class CreateSavingGoalRequest {
     private int user_id;
     private String name;
-    private double current_amount;
-    private double target_amount;
+    private BigDecimal current_amount;
+    private BigDecimal target_amount;
     private String description;
     private String start_date;
     private String target_date;
     private int category_id;
 
-    public CreateSavingGoalRequest(int user_id, String name, double amount, double target,
+    public CreateSavingGoalRequest(int user_id, String name, BigDecimal amount, BigDecimal target,
                                    String description, String start_date, String end_date, int category_id) {
         this.user_id = user_id;
         this.name = name;
@@ -24,8 +26,8 @@ public class CreateSavingGoalRequest {
 
     public int getUser_id() { return user_id; }
     public String getName() { return name; }
-    public double getCurrent_amount() { return current_amount; }
-    public double getTarget_amount() { return target_amount; }
+    public BigDecimal getCurrent_amount() { return current_amount; }
+    public BigDecimal getTarget_amount() { return target_amount; }
     public String getDescription() { return description; }
     public String getStart_date() { return start_date; }
     public String getTarget_date() { return target_date; }
